backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
opencv-python-headless==4.10.0.84
websockets==12.0
ultralytics==8.3.0
google-genai
python-dotenv
datasets
